Adding new shared components
