+++ src/components/shared/index.js
@@ -1,2 +1,4 @@
 // Export all shared components for easier imports
 export { default as ChatInterface } from './ChatInterface';
+export { default as PageHeader } from './PageHeader';
+export { default as StatCard } from './StatCard';
